Add tests for Entry and Logger logging methods
After processing of a log message is completed, testing of all logging
methods becomes available.

Both Entry and Logger methods are tested using backend with mocked up
filter, serializer and writer.

In all tests threshold is set to Debug level, so that all messages
logged during the tests pass threshold and are passed to be processed
by backends.

Change-Id: I10a1365b255fb0caf083f5dee19de889d92e42a5
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>