Implement writer to file
The WriterFile type implements Writer interface.
It writes received data to the file, which is open during
creation of the WriterFile object with NewWriterFile function.
Every log message is written to separate line.

If file does not exist, it is created.
If file exists it is opened in append mode.

NewWriteFile function panics, if it cannot open the file.

Change-Id: I553cd042d3a873f2d6c7b1e33861f8c1ae74bc13
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>