Fix SerializerText timestamp tests
The time variables used for tests were defined in local timezone.
On different machines the output of timestamp serialization was also
in different timezones.
This was fixed to use UTC timezone for that test.

To cover also serilization in different timezones an additional test
was added to verify serialization in different timezones.

Change-Id: I69e9530006c425426f944d2591c8ae4a5e3713ce
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>